+++ packages/explorer/src/domain/objects/entities/index.ts
@@ -1,5 +1,4 @@
 export * from "./object";
-export * from "./objectList";
 export * from "./filters";
 export * from "./mjdUtils";
 export * from "./coordUtils";