+++ packages/explorer/src/ui/stores/search/di.ts
@@ -1,4 +1,5 @@
 import {
+  convertGregUseCase,
   convertMjdUseCase,
   getObjectListUseCase,
 } from "@/app/object/use-cases";
@@ -6,7 +7,8 @@ import { searchStore } from "./store";
 
 const searchStoreResolved = searchStore(
   getObjectListUseCase,
-  convertMjdUseCase
+  convertMjdUseCase,
+  convertGregUseCase
 );
 
 export { searchStoreResolved as useSearchStore };