+++ packages/explorer/src/ui/stores/search/store.ts
@@ -1,16 +1,18 @@
 import type { Command } from "@/common/use-case";
 import type {
-  ObjectListEntity,
+  ObjectEntity,
   ObjectListFilters,
 } from "@/domain/objects/entities";
 import type {
   HttpError,
+  PaginatedListEntity,
   ParseError,
 } from "@alercebroker/http-client/build/main/types";
 import { defineStore } from "pinia";
 import { computed, ref } from "vue";
 import { parseInput, parseInputReverse } from "./parseInput";
 import type { SearchInput } from "./types";
+import { validateInputFilters } from "./validateInput";
 
 export type PremadeQuery = {
   title: string;
@@ -25,11 +27,13 @@ export type Errors = {
   client: HttpError | null;
   server: HttpError | null;
   parse: ParseError | null;
+  inputError: Error[] | null;
 };
 
 export const searchStore = (
   searchObjectsUseCase: Command,
-  convertMjdUseCase: Command
+  convertMjdUseCase: Command,
+  convertGregUseCase: Command
 ) => {
   return defineStore("search", () => {
     const filters = ref<ObjectListFilters>({
@@ -41,6 +45,7 @@ export const searchStore = (
       dec: -999,
       radius: -999,
     });
+
     const premadeQueries = ref<PremadeQuery[]>([
       {
         title: "Query Title 1",
@@ -75,7 +80,8 @@ export const searchStore = (
         filters: {},
       },
     ]);
-    const results = ref<ObjectListEntity>({
+
+    const results = ref<PaginatedListEntity<ObjectEntity>>({
       total: 0,
       page: 1,
       next: 2,
@@ -84,36 +90,83 @@ export const searchStore = (
       hasPrev: false,
       items: [],
     });
+
     const errors = ref<Errors>({
       generic: null,
       client: null,
       server: null,
       parse: null,
+      inputError: null,
     });
 
     function search(inputFilters: SearchInput) {
+      const [isValid, inputErrors] = validateInputFilters(inputFilters);
+      if (!isValid) {
+        errors.value.inputError = inputErrors.map((errorsito) => {
+          return new Error(errorsito);
+        });
+        return;
+      }
       const parsedFilters = parseInput(inputFilters);
       filters.value = parsedFilters;
       searchObjectsUseCase.execute(
         {
-          handleSuccess: (data: ObjectListEntity) => {
+          handleSuccess: (data: PaginatedListEntity<ObjectEntity>) => {
             results.value = data;
           },
-          handleGenericError: (error: Error) => {
-            errors.value.generic = error;
-          },
-          handleHttpClientError: (error: HttpError) => {
-            errors.value.client = error;
-          },
-          handleHttpServerError: (error: HttpError) => {
-            errors.value.server = error;
-          },
-          handleParseError: (error: ParseError) => {
-            errors.value.parse = error;
+          handleError: {
+            handleGenericError: (error: Error) => {
+              errors.value.generic = error;
+            },
+            handleHttpClientError: (error: HttpError) => {
+              errors.value.client = error;
+            },
+            handleHttpServerError: (error: HttpError) => {
+              errors.value.server = error;
+            },
+            handleParseError: (error: ParseError) => {
+              errors.value.parse = error;
+            },
           },
         },
         parsedFilters
       );
+    }
+
+    function convertGregToMjd(gregDate: string): number {
+      let result = -999;
+      convertGregUseCase.execute(
+        {
+          handleSuccess: (mjd: number) => {
+            result = mjd;
+          },
+          handleError: {
+            handleGenericError: (error) => {
+              errors.value.generic = error;
+            },
+          },
+        },
+        gregDate
+      );
+      return result;
+    }
+
+    function convertMjdToGreg(mjd: number): string {
+      let result = "";
+      convertMjdUseCase.execute(
+        {
+          handleSuccess: (greg: string) => {
+            result = greg;
+          },
+          handleError: {
+            handleGenericError: (error) => {
+              errors.value.generic = error;
+            },
+          },
+        },
+        mjd
+      );
+      return result;
     }
 
     const componentFilters = computed(() => {
@@ -127,6 +180,8 @@ export const searchStore = (
       search,
       componentFilters,
       results,
+      convertGregToMjd,
+      convertMjdToGreg,
     };
   });
 };