+++ packages/hunter/src/ui/stores/di.ts
@@ -1,0 +1,5 @@
+import { objectStoreFactory } from "./objects";
+import { getObjectList } from "@/application/objects";
+
+const objectStore = objectStoreFactory(getObjectList);
+export { objectStore as useObjectStore };